Add files via upload
Using the four pillars of Object-Oriented Programming (inheritance, polymorphism, abstraction, and encapsulation), provides you with different information according to the device (television, videogame console, car, supercomputer). To make the code shorter to write and easier to understand,  we executed parent class and child classes. In addition, getters and setters are used in the code so that it is possible to read and write (respectively) the values ​​of our private variables from outside the class where they were created