Add files via upload
The following code is an example of how constructors are used in C++. To customize how a class initializes its members or to invoke functions when an object of the class is created, define a constructor. A constructor has the same name as the class and does not return a value.